+++ src/components/details-page/details-page.test.js
@@ -3,7 +3,7 @@ import {shallow} from 'enzyme';
 
 import {DetailsPage} from '../details-page/details-page.jsx';
 import offers from '../../mocks/offers.js';
-import reviews from '../../mocks/reviews.js';
+import comments from '../../mocks/comments.js';
 
 const city = offers[0].city;
 
@@ -11,11 +11,11 @@ it(`DetailsPage renders correctly`, () => {
   const tree = shallow(
       <DetailsPage
         offers={offers}
-        reviews={reviews}
+        comments={comments}
         city={city}
         onHoverOfferId={-1}
-        getOfferId={() => {}}
-        fetchComments={() => {}}
+        getOfferIdHandler={() => {}}
+        fetchCommentsHandler={() => {}}
       />
   );
