+++ src/components/reviews-item/reviews-item.test.js
@@ -2,10 +2,10 @@ import React from 'react';
 import {shallow} from 'enzyme';
 
 import ReviewsItem from '../reviews-item/reviews-item.jsx';
-import reviews from '../../mocks/reviews.js';
+import comments from '../../mocks/comments.js';
 
 it(`ReviewsItem renders correctly`, () => {
-  const tree = shallow(<ReviewsItem review={reviews[0]} key={reviews[0].id}/>);
+  const tree = shallow(<ReviewsItem review={comments[0]} key={comments[0].id}/>);
 
   expect(tree).toMatchSnapshot();
 });
