+++ src/components/reviews-list/reviews-list.test.js
@@ -2,10 +2,10 @@ import React from 'react';
 import {shallow} from 'enzyme';
 
 import ReviewsList from '../reviews-list/reviews-list.jsx';
-import reviews from '../../mocks/reviews.js';
+import comments from '../../mocks/comments.js';
 
 it(`ReviewsList renders correctly`, () => {
-  const tree = shallow(<ReviewsList reviews={reviews} />);
+  const tree = shallow(<ReviewsList reviews={comments} />);
 
   expect(tree).toMatchSnapshot();
 });
